+++ effects/flame.h
@@ -27,7 +27,7 @@ class FlameEffect
         virtual void windowClosed( EffectWindow* c );
         virtual void windowDeleted( EffectWindow* c );
     private:
-        QHash< const EffectWindow*, double > windows;
+        QHash< const EffectWindow*, float > windows;
     };
 
 } // namespace
